requirements.txt
Django~=5.2.7
djangorestframework~=3.16.1
python-dotenv~=1.2.1
django-filter~=25.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

